requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
urllib3==2.3.0
